reorg, add actions
