+++ .github/workflows/node.js.yml
@@ -1,0 +1,47 @@
+# This workflow will do a clean install of node dependencies, run the dockerimage, checkout the basebot code, and execute the tests across different versions of node
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
+
+name: Node.js CI
+
+on:
+  push:
+    # branches: [master]
+    branches: [babelDMonly]
+  pull_request:
+    # branches: [master]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    continue-on-error: ${{ matrix.experimental }}
+    strategy:
+      matrix:
+        node-version: [10.x]
+        experimental: [false]
+        include:
+          - node: 12
+            experimental: true
+        # - node: 14
+        #   experimental: true
+    steps:
+      - name: Login to Docker
+        uses: azure/docker-login@v2
+        with:
+          username: ${{ secrets.DOCKER_USERNAME }}
+          password: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Run docker image
+      # run: docker run -v ~/opt/WickrIO:/opt/WickrIODebug -d -p 4545:4545 -p 9229:9229 --restart=always --name="WickrBotAPIContainer" -ti wickr/bot-cloud-alpha:5.68.01.08
+      run: docker run -v ~/opt/WickrIO:/opt/WickrIODebug -d --restart=always --name="WickrBotAPIContainer" -ti wickr/bot-cloud-alpha:5.68.01.08
+
+      - uses: actions/checkout@v2
+      - name: Checkout, Test Using Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ matrix.node-version }}
+      - run: npm install
+      - run: npm run build --if-present
+      # - run: npm run test
+        # env:
+        #   tokens: ''
