+++ .gitignore
@@ -3,4 +3,3 @@ package-lock.json
 /test
 /lib
 build
-.github
