+++ src/api/index.js
@@ -1,0 +1,7 @@
+import BotAPI from './BotAPI'
+import ConfigBot from './ConfigBot'
+import WickrAdmin from './WickrAdmin'
+import WickrStrings from './WickrStrings'
+import WickrUser from './WickrUser'
+
+export { BotAPI, ConfigBot, WickrAdmin, WickrStrings, WickrUser }
