+++ src/services/index.js
@@ -1,0 +1,3 @@
+import MessageService from './message'
+import APIService from './wickrapi'
+export { MessageService, APIService }
